% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_draws.R
\name{collect_draws}
\alias{collect_draws}
\title{Scrape Draw Results}
\usage{
collect_draws(urls)
}
\arguments{
\item{urls.}{Vector of Grand Slam Wiki event urls}
}
\value{
Data frame with game score, players, round and event for each Grand Slam
}
\description{
Scrape Draw Results
}
\details{
The \code{tiebreak} column are the points won in the tiebreak, if played
}
