% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_players.R
\name{collect_players}
\alias{collect_players}
\title{Scapre Table of Players with Wiki Pages}
\usage{
collect_players(urls)
}
\arguments{
\item{urls.}{Vector of character of event urls}
}
\value{
Data frame of \code{player_page} URL to uniquely identify a player and \code{player} (which may have multiple variants)
}
\description{
Scapre Table of Players with Wiki Pages
}
