% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_draw.R
\name{read_draw}
\alias{read_draw}
\title{Read Tables of Draw Results and Combine}
\usage{
read_draw(url)
}
\arguments{
\item{url.}{A url of the grand slam singles event}
}
\value{
a data.frame of the match results for the event
}
\description{
Read Tables of Draw Results and Combine
}
