% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_events.R
\name{read_events}
\alias{read_events}
\title{Scrape Wiki URLS For Events}
\usage{
read_events(
  url = "https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Tennis/Grand_Slam_Project"
)
}
\arguments{
\item{url.}{A character of the wiki page with Grand Slam event URLS}
}
\value{
Data frame of \code{links} that uniquely identify and event descriptors
}
\description{
Scrape Wiki URLS For Events
}
