% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_irregularities.R
\name{read_irregularities}
\alias{read_irregularities}
\title{Capture Tables That Don't Follow Knockout Format}
\usage{
read_irregularities(url)
}
\arguments{
\item{url.}{A character URL}
}
\value{
A data frame of each table structure and any that don't follow expected power of 2 knockout structure
}
\description{
Capture Tables That Don't Follow Knockout Format
}
