% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_round_table.R
\name{read_round_table}
\alias{read_round_table}
\title{Read Rounds of Draw}
\usage{
read_round_table(table, tablenum, url)
}
\arguments{
\item{table.}{A table node from \code{html_nodes}}

\item{tablenum.}{A numeric number for the index of the table in the page's html tables}

\item{url.}{A character url}
}
\value{
A data frame of match results by round
}
\description{
Read Rounds of Draw
}
