% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wimbledon-players-data.R
\docType{data}
\name{wimbledon_players}
\alias{wimbledon_players}
\title{Wimbeldon Players}
\format{A data.frame with the following variables:

\describe{
  \item{player_page}{Character URL of player page (if it exists) on Wikipedia. Add this to wikipedia.org for complete URL.}
  \item{player}{Character name of player. There may be multiple variables per player}
}}
\source{
\href{https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Tennis/Grand_Slam_Project}{Wiki Grand Slam Project}
}
\usage{
data(wimbledon_players)
}
\description{
Data of player names and player pages on Wikipedia
}
\keyword{datasets}
